1. Cleanup Project 2. Change the Java Compiler version from 1.6 to 1.7 3. Change installed Java facet version from 1.6 to 1.7

+++ src/com/cqut/xiji/controller/abilityCheck/AbilityCheckController.java
@@ -1,16 +1,11 @@
 package com.cqut.xiji.controller.abilityCheck;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.UUID;
-
 import javax.annotation.Resource;
 
 
@@ -25,16 +20,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
 import com.cqut.xiji.dao.base.EntityDao;
 import com.cqut.xiji.entity.abilityCheck.AbilityCheck;
 import com.cqut.xiji.entity.fileInformation.FileInformation;
 import com.cqut.xiji.service.abilityCheck.IAbilityCheckService;
-import com.cqut.xiji.service.fileOperate.IFileOperateService;
-import com.cqut.xiji.tool.util.EntityIDFactory;
 
 @Controller
 @RequestMapping("/abilityCheckController")
@@ -179,7 +170,7 @@ public class AbilityCheckController{
 			String ID = request.getParameter("fileID");// 文件ID
 			String fileName = file.getOriginalFilename();// 文件名
 			String path = "";// 文件路径
-			begin = (long) System.currentTimeMillis();
+			begin = System.currentTimeMillis();
 			fileNames = fileName.split("\\.");
 			
 			path = System.getProperty("user.dir").replace("bin", "webapps")  + "\\"
